API_VIEWSET/viewset_app/requirement.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.0.5
djangorestframework==3.12.4
djongo==1.3.4
idna==2.10
pymongo==3.11.4
pytz==2021.1
requests==2.25.1
sqlparse==0.2.4
urllib3==1.26.5

using by default data sqlite3
username=Anshu
password=@Automata102
